api/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
certifi==2024.2.2
click==8.1.7
deprecation==2.1.0
fastapi==0.110.0
gotrue==2.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
idna==3.6
packaging==24.0
postgrest==0.16.1
psycopg2==2.9.9
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.2
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
storage3==0.7.3
StrEnum==0.4.15
supabase==2.4.0
supafunc==0.3.3
typing_extensions==4.10.0
uvicorn==0.28.0
websockets==11.0.3
